+++ src/main/java/racingcar/application/implement/WinnerIdentifier.java
@@ -1,22 +1,8 @@
 package racingcar.application.implement;
 
 import java.util.List;
-import racingcar.domain.CarRacer;
 
-public class WinnerIdentifier {
+public interface WinnerIdentifier<T> {
 
-    public List<CarRacer> identify(List<CarRacer> racedCars) {
-        int maxProgress = getMaxProgress(racedCars);
-
-        return racedCars.stream()
-                .filter(car -> car.getProgressState().length() == maxProgress)
-                .toList();
-    }
-
-    private int getMaxProgress(List<CarRacer> racedCars) {
-        return racedCars.stream()
-                .mapToInt(car -> car.getProgressState().length())
-                .max()
-                .getAsInt();
-    }
+    List<T> identify(List<T> participants);
 }
